Extract estimated parameters from a fit_gnss_ts_ngl

Extract estimated parameters from a fit_gnss_ts_ngl

# S3 method for class 'fit_gnss_ts_ngl'
summary(object, scale_parameters = FALSE, ...)

Arguments

object

A fit_gnss_ts_ngl object.

scale_parameters

A boolean indicating whether or not to scale estimated parameters so that the returned estimated trend is provided in m/year instead of m/day. Default is FALSE.

...

Additional parameters.

Examples

x <- download_station_ngl("P820")
fit1 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + pl")
summary(fit1)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.74690707 389.79832189 -763.24376507 764.73757921
#> Trend                   -0.00000721   0.00342891  -0.00672775   0.00671333
#> Sin (Annual)            -0.00273991   3.56321214  -6.98650737   6.98102756
#> Cos (Annual)            -0.00277131   3.46650769  -6.79700154   6.79145892
#> Sin (Semi-Annual)        0.00087464   3.35377338  -6.57240040   6.57414968
#> Cos (Semi-Annual)       -0.00059571   3.34624511  -6.55911560   6.55792418
#> Jump: MJD 60649         -0.00010784  16.51627880 -32.37141944  32.37120376
#> Earthquake: MJD 60649    0.00072044  37.03198435 -72.58063516  72.58207605
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000058
#>  Stationary powerlaw Spectral index:    -1.00000000
#>  Stationary powerlaw Variance:     0.00000078
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00065789
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.10000000
#>  \hat{E[Z]}: 0.99346405
#> -------------------------------------------------------------
#> Running time: 0.47 seconds
#> -------------------------------------------------------------
summary(fit1, scale_parameters = TRUE)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept              272.80780701 142373.83706985 -278774.78519067 279320.40080468
#> Trend                   -0.00263347   1.25240917  -2.45731034   2.45204340
#> Sin (Annual)            -1.00075049 1301.46323398 -2551.82181629 2549.82031532
#> Cos (Annual)            -1.01222083 1266.14193492 -2482.60481258 2480.58037092
#> Sin (Semi-Annual)        0.31946249 1224.96572695 -2400.56924463 2401.20816961
#> Cos (Semi-Annual)       -0.21758349 1222.21602545 -2395.71697469 2395.28180772
#> Jump: MJD 60649         -0.03938892 6032.57083058 -11823.66095104 11823.58217319
#> Earthquake: MJD 60649    0.26314229 13525.93228495 -26510.07699354 26510.60327811
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000058
#>  Stationary powerlaw Spectral index:    -1.00000000
#>  Stationary powerlaw Variance:     0.00000078
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00065789
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.10000000
#>  \hat{E[Z]}: 0.99346405
#> -------------------------------------------------------------
#> Running time: 0.47 seconds
#> -------------------------------------------------------------
fit2 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + fl")
summary(fit2)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.74690707   0.00073637   0.74546382   0.74835032
#> Trend                   -0.00000721   0.00000052  -0.00000823  -0.00000619
#> Sin (Annual)            -0.00273991   0.00017680  -0.00308643  -0.00239338
#> Cos (Annual)            -0.00277131   0.00017409  -0.00311252  -0.00243010
#> Sin (Semi-Annual)        0.00087464   0.00012456   0.00063051   0.00111878
#> Cos (Semi-Annual)       -0.00059571   0.00012330  -0.00083738  -0.00035404
#> Jump: MJD 60649         -0.00010784   0.00117394  -0.00240872   0.00219304
#> Earthquake: MJD 60649    0.00072044   0.00238875  -0.00396142   0.00540231
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000058
#>  Flicker Noise Variance:     0.00000078
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00065789
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.10000000
#>  \hat{E[Z]}: 0.99346405
#> -------------------------------------------------------------
#> Running time: 0.92 seconds
#> -------------------------------------------------------------