Extract estimated parameters from a fit_gnss_ts_ngl

Extract estimated parameters from a fit_gnss_ts_ngl

Usage

# S3 method for class 'fit_gnss_ts_ngl'
summary(object, scale_parameters = FALSE, ...)

Arguments

object

A fit_gnss_ts_ngl object.

scale_parameters

A boolean indicating whether or not to scale estimated parameters so that the returned estimated trend is provided in m/year instead of m/day. Default is FALSE.

...

Additional parameters.

Examples

x <- download_station_ngl("P820")
fit1 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + pl")
summary(fit1)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.74683110   0.00002216   0.74678767   0.74687453
#> Trend                   -0.00000721   0.00000003  -0.00000726  -0.00000716
#> Sin (Annual)            -0.00274600   0.00002887  -0.00280257  -0.00268942
#> Cos (Annual)            -0.00277284   0.00002826  -0.00282823  -0.00271745
#> Sin (Semi-Annual)        0.00086979   0.00002823   0.00081446   0.00092511
#> Cos (Semi-Annual)       -0.00059347   0.00002830  -0.00064894  -0.00053801
#> Jump: MJD 60649         -0.00006729   0.00013537  -0.00033260   0.00019803
#> Earthquake: MJD 60649    0.00057701   0.00029152   0.00000564   0.00114838
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000241
#>  Stationary powerlaw Spectral index:     0.99998941
#>  Stationary powerlaw Variance:     0.00000000
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00065338
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.10000000
#>  \hat{E[Z]}: 0.99350860
#> -------------------------------------------------------------
#> Running time: 0.27 seconds
#> -------------------------------------------------------------
summary(fit1, scale_parameters = TRUE)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept              272.78005948   0.00809404 272.76419545 272.79592351
#> Trend                   -0.00263361   0.00001014  -0.00265348  -0.00261374
#> Sin (Annual)            -1.00297488   0.01054307  -1.02363891  -0.98231084
#> Cos (Annual)            -1.01277998   0.01032306  -1.03301280  -0.99254716
#> Sin (Semi-Annual)        0.31768952   0.01030988   0.29748254   0.33789651
#> Cos (Semi-Annual)       -0.21676614   0.01033681  -0.23702591  -0.19650637
#> Jump: MJD 60649         -0.02457606   0.04944323  -0.12148301   0.07233088
#> Earthquake: MJD 60649    0.21075279   0.10647764   0.00206046   0.41944512
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000241
#>  Stationary powerlaw Spectral index:     0.99998941
#>  Stationary powerlaw Variance:     0.00000000
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00065338
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.10000000
#>  \hat{E[Z]}: 0.99350860
#> -------------------------------------------------------------
#> Running time: 0.27 seconds
#> -------------------------------------------------------------
fit2 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + fl")
summary(fit2)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.74683110   0.00074186   0.74537708   0.74828512
#> Trend                   -0.00000721   0.00000052  -0.00000824  -0.00000618
#> Sin (Annual)            -0.00274600   0.00017546  -0.00308989  -0.00240210
#> Cos (Annual)            -0.00277284   0.00017494  -0.00311572  -0.00242996
#> Sin (Semi-Annual)        0.00086979   0.00012450   0.00062578   0.00111380
#> Cos (Semi-Annual)       -0.00059347   0.00012314  -0.00083483  -0.00035211
#> Jump: MJD 60649         -0.00006729   0.00116972  -0.00235990   0.00222533
#> Earthquake: MJD 60649    0.00057701   0.00231557  -0.00396142   0.00511544
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000057
#>  Flicker Noise Variance:     0.00000079
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00065338
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.10000000
#>  \hat{E[Z]}: 0.99350860
#> -------------------------------------------------------------
#> Running time: 0.94 seconds
#> -------------------------------------------------------------