Extract estimated parameters from a fit_gnss_ts_ngl

Extract estimated parameters from a fit_gnss_ts_ngl

# S3 method for class 'fit_gnss_ts_ngl'
summary(object, scale_parameters = FALSE, ...)

Arguments

object

A fit_gnss_ts_ngl object.

scale_parameters

A boolean indicating whether or not to scale estimated parameters so that the returned estimated trend is provided in m/year instead of m/day. Default is FALSE.

...

Additional parameters.

Examples

x <- download_station_ngl("P820")
fit1 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + pl")
summary(fit1)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.75007627   0.00405363   0.74213130   0.75802125
#> Trend                   -0.00000700   0.00000060  -0.00000818  -0.00000582
#> Sin (Annual)             0.00065832   0.00019794   0.00027035   0.00104628
#> Cos (Annual)            -0.00050112   0.00020731  -0.00090744  -0.00009480
#> Sin (Semi-Annual)        0.00044663   0.00014636   0.00015977   0.00073349
#> Cos (Semi-Annual)        0.00012070   0.00014257  -0.00015874   0.00040014
#> Jump: MJD 60547          0.00229644   0.00118728  -0.00003059   0.00462347
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000091
#>  Stationary powerlaw Spectral index:    -0.95555862
#>  Stationary powerlaw Variance:     0.00000099
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00076834
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 1.00000000
#>  \hat{E[Z]}: 0.99923225
#> -------------------------------------------------------------
#> Running time: 0.51 seconds
#> -------------------------------------------------------------
summary(fit1, scale_parameters=TRUE)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept              273.96535795   1.48058958 271.06345569 276.86726021
#> Trend                   -0.00255669   0.00022017  -0.00298822  -0.00212517
#> Sin (Annual)             0.24045103   0.07229927   0.09874706   0.38215501
#> Cos (Annual)            -0.18303481   0.07572044  -0.33144414  -0.03462547
#> Sin (Semi-Annual)        0.16313051   0.05345811   0.05835454   0.26790647
#> Cos (Semi-Annual)        0.04408635   0.05207533  -0.05797942   0.14615212
#> Jump: MJD 60547          0.83877580   0.43365496  -0.01117230   1.68872389
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000091
#>  Stationary powerlaw Spectral index:    -0.95555862
#>  Stationary powerlaw Variance:     0.00000099
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00076834
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 1.00000000
#>  \hat{E[Z]}: 0.99923225
#> -------------------------------------------------------------
#> Running time: 0.51 seconds
#> -------------------------------------------------------------
fit2 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + fl")
summary(fit2)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.75007627   0.00075821   0.74859022   0.75156233
#> Trend                   -0.00000700   0.00000059  -0.00000816  -0.00000584
#> Sin (Annual)             0.00065832   0.00020072   0.00026491   0.00105173
#> Cos (Annual)            -0.00050112   0.00021071  -0.00091410  -0.00008814
#> Sin (Semi-Annual)        0.00044663   0.00014791   0.00015672   0.00073653
#> Cos (Semi-Annual)        0.00012070   0.00014389  -0.00016132   0.00040273
#> Jump: MJD 60547          0.00229644   0.00121168  -0.00007841   0.00467129
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000098
#>  Flicker Noise Variance:     0.00000091
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00076834
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 1.00000000
#>  \hat{E[Z]}: 0.99923225
#> -------------------------------------------------------------
#> Running time: 0.74 seconds
#> -------------------------------------------------------------