Extract estimated parameters from a fit_gnss_ts_ngl

Extract estimated parameters from a fit_gnss_ts_ngl

# S3 method for class 'fit_gnss_ts_ngl'
summary(object, scale_parameters = FALSE, ...)

Arguments

object

A fit_gnss_ts_ngl object.

scale_parameters

A boolean indicating whether or not to scale estimated parameters so that the returned estimated trend is provided in m/year instead of m/day. Default is FALSE.

...

Additional parameters.

Examples

x <- download_station_ngl("P820")
fit1 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + pl")
summary(fit1)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.75010524   0.00560599   0.73911770   0.76109277
#> Trend                   -0.00000693   0.00000062  -0.00000814  -0.00000572
#> Sin (Annual)             0.00060825   0.00020361   0.00020919   0.00100732
#> Cos (Annual)            -0.00047066   0.00021055  -0.00088333  -0.00005800
#> Sin (Semi-Annual)        0.00040375   0.00014749   0.00011468   0.00069283
#> Cos (Semi-Annual)        0.00008541   0.00014720  -0.00020311   0.00037393
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000088
#>  Stationary powerlaw Spectral index:    -0.95767473
#>  Stationary powerlaw Variance:     0.00000103
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00076834
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 1.00000000
#>  \hat{E[Z]}: 0.99923225
#> -------------------------------------------------------------
#> Running time: 0.36 seconds
#> -------------------------------------------------------------
summary(fit1, scale_parameters = TRUE)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept              273.97593721   2.04758744 269.96273957 277.98913484
#> Trend                   -0.00253039   0.00022526  -0.00297188  -0.00208890
#> Sin (Annual)             0.22216452   0.07436781   0.07640630   0.36792275
#> Cos (Annual)            -0.17190992   0.07690285  -0.32263674  -0.02118310
#> Sin (Semi-Annual)        0.14747050   0.05387064   0.04188598   0.25305502
#> Cos (Semi-Annual)        0.03119569   0.05376653  -0.07418477   0.13657615
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000088
#>  Stationary powerlaw Spectral index:    -0.95767473
#>  Stationary powerlaw Variance:     0.00000103
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00076834
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 1.00000000
#>  \hat{E[Z]}: 0.99923225
#> -------------------------------------------------------------
#> Running time: 0.36 seconds
#> -------------------------------------------------------------
fit2 <- gmwmx2(x, n_seasonal = 2, component = "N", stochastic_model = "wn + fl")
summary(fit2)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept                0.75010524   0.00077387   0.74858849   0.75162198
#> Trend                   -0.00000693   0.00000060  -0.00000810  -0.00000576
#> Sin (Annual)             0.00060825   0.00020543   0.00020561   0.00101089
#> Cos (Annual)            -0.00047066   0.00021283  -0.00088780  -0.00005353
#> Sin (Semi-Annual)        0.00040375   0.00014855   0.00011260   0.00069490
#> Cos (Semi-Annual)        0.00008541   0.00014818  -0.00020502   0.00037584
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000095
#>  Flicker Noise Variance:     0.00000095
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00076834
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 1.00000000
#>  \hat{E[Z]}: 0.99923225
#> -------------------------------------------------------------
#> Running time: 0.69 seconds
#> -------------------------------------------------------------