Extract estimated parameters from a fit_gnss_ts_ngl

Extract estimated parameters from a fit_gnss_ts_ngl

# S3 method for class 'fit_gnss_ts_ngl'
summary(object, scale_parameters = FALSE, ...)

Arguments

object

A fit_gnss_ts_ngl object.

scale_parameters

A boolean indicating whether or not to scale estimated parameters so that the returned estimated trend is provided in m/year instead of m/day. Default is FALSE.

...

Additional parameters.

Examples

x <- download_station_ngl("CHML")
fit <- gmwmx2(x, n_seasonal = 2, component = "N")
summary(fit)
#> Summary of Estimated Model
#> -------------------------------------------------------------
#> Functional parameters
#> -------------------------------------------------------------
#> Parameter                  Estimate  Std_Deviation  95% CI Lower  95% CI Upper
#> -------------------------------------------------------------
#> Intercept               -0.70924596   0.01136945  -0.73152967  -0.68696225
#> Trend                    0.00007198   0.00001083   0.00005074   0.00009321
#> Sin (Annual)             0.00135377   0.00043414   0.00050287   0.00220467
#> Cos (Annual)            -0.00062838   0.00045622  -0.00152254   0.00026579
#> Sin (Semi-Annual)        0.00017180   0.00027821  -0.00037348   0.00071707
#> Cos (Semi-Annual)        0.00030877   0.00028861  -0.00025689   0.00087443
#> Jump: MJD 56248          0.01588617   0.00191916   0.01212469   0.01964764
#> Jump: MJD 55391          0.00780478   0.00182956   0.00421891   0.01139065
#> Jump: MJD 55563          0.00230763   0.00190189  -0.00142001   0.00603527
#> Jump: MJD 55603          0.00286880   0.00235741  -0.00175163   0.00748923
#> Jump: MJD 55606         -0.00044386   0.00358112  -0.00746272   0.00657500
#> Jump: MJD 56011         -0.00008945   0.00192695  -0.00386620   0.00368730
#> Jump: MJD 56085          0.00073436   0.00199984  -0.00318524   0.00465397
#> Jump: MJD 56748         -0.00038578   0.00187908  -0.00406870   0.00329714
#> Jump: MJD 57281         -0.00145742   0.00187987  -0.00514189   0.00222704
#> Earthquake: MJD 55391    0.01576163   0.00930876  -0.00248320   0.03400647
#> Earthquake: MJD 55563    0.00174787   0.02766816  -0.05248072   0.05597647
#> Earthquake: MJD 55603   -0.28371260   0.52755951  -1.31771023   0.75028503
#> Earthquake: MJD 55606    0.27719703   0.52174624  -0.74540681   1.29980088
#> Earthquake: MJD 56011   -0.00480858   0.01509158  -0.03438754   0.02477039
#> Earthquake: MJD 56085   -0.00772227   0.01454627  -0.03623244   0.02078791
#> Earthquake: MJD 56748   -0.00910404   0.00725836  -0.02333016   0.00512207
#> Earthquake: MJD 57281   -0.02236505   0.00811536  -0.03827086  -0.00645925
#> -------------------------------------------------------------
#> Stochastic parameters
#> -------------------------------------------------------------
#>  White Noise Variance  :     0.00000193
#>  Flicker Noise Variance:     0.00000251
#> -------------------------------------------------------------
#> Missingness parameters
#> -------------------------------------------------------------
#>  P(Z_{i+1} = 0 | Z_{i} = 1): 0.00279460
#>  P(Z_{i+1} = 1 | Z_{i} = 0): 0.31578947
#>  \hat{E[Z]}: 0.99122807
#> -------------------------------------------------------------
#> Running time: 0.8 seconds
#> -------------------------------------------------------------